+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 ### Added
+- Added updated field to task
+- Added update operation (missing comments)
 - Added delete operation
 
 ## [0.0.2] 2020-08-30