+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Refactored task manager
+
+## [0.0.3] 2020-09-05
+### Added
+- Added updated field to task
+- Added update operation (missing comments)
+- Added delete operation
+
+## [Unreleased]
 ### Added
 - Added updated field to task
 - Added update operation (missing comments)